[REPLAY] Extract shadow DOM logic in its file
